+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- If a local migration with the same version already exists, bolt will now error out when listing out the migrations and indicate that there is a conflicting migration. Previously, it would ignore the first migration and use the second one with the same version.
+
 ## [0.8.0] - 2024-03-16
 
 ### Changed